Setup ALSA for Linux build in .travis.yml

+++ scripts/install-alsa.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-sudo apt-get remove --purge alsa-base pulseaudio
-sudo apt-get install alsa-base pulseaudio
-sudo alsa force-reload
-sudo apt install libasound2-dev
+
+if [[ $TRAVIS_OS_NAME == 'linux' ]]; then
+  sudo apt-get remove --purge alsa-base pulseaudio
+  sudo apt-get install alsa-base pulseaudio
+  sudo alsa force-reload
+  sudo apt install libasound2-dev
+fi
